requirements.txt
streamlit == 1.23.0
tensorflow == 2.8.0
transformers == 4.29.2
tweet-preprocessor == 0.6.0